+++ index.ts
@@ -2,10 +2,11 @@ export * from "./src/decorators/method_decorator";
 export * from "./src/decorators/method_paramter_decorators";
 export * from "./src/decorators/default_handler";
 export { extender, BeforeCreate, SetProperty, AfterCreate } from "./src/util/class_decorator_util";
-export { App } from "./src/app";
-export { Router } from "./src/router";
+export * from "./src/classes/app";
+export * from "./src/classes/middleware";
+export * from "./src/classes/router";
 export * from "./src/decorators/class_decorator";
-export * from "./src/middleware";
+export * from "./src/decorators/middleware_decorators";
 
 import { Response, Request, ParamsDictionary, Router as ExpressRouter } from "express-serve-static-core";
 import { ParsedQs } from 'qs';
