+++ src/classes/app.ts
@@ -1,6 +1,6 @@
 import express, { Application } from "express";
 import { Server } from "node:http";
-import { Routable } from "./classes/routable";
+import { Routable } from "./routable";
 
 export abstract class App extends Routable<Application> {
   public path: string = '/';
@@ -14,15 +14,15 @@ export abstract class App extends Routable<Application> {
     this.running_server = undefined;
   }
 
-  set start_stop_logging (value: boolean) {
+  public set start_stop_logging (value: boolean) {
     this._start_stop_logging = value;
   }
 
-  remove_layers(): void {
+  public remove_layers(): void {
     this.get_routable()._router.stack.splice(2);
   }
 
-  start_app () {
+  public start_app () {
     if (!this.layers_initialized) {
       this.setup_layers();
     }
@@ -37,7 +37,7 @@ export abstract class App extends Routable<Application> {
       });
   }
 
-  stop (on_stop_callback: () => any = () => {}) {
+  public stop (on_stop_callback: () => any = () => {}) {
     if (this.running_server) {
       this.running_server.close(() => {
         delete this.running_server;
