+++ src/classes/router.ts
@@ -1,9 +1,7 @@
-import {Router as ExpressRouter} from "express";
-import {Routable} from "./classes/routable";
+import { Router as ExpressRouter } from "express";
+import { Routable } from "./routable";
 
 export class Router extends Routable<ExpressRouter> {
-  // public static path: string = '/';
-
   constructor() {
     super(ExpressRouter());
   }
